requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
